+++ src/main/java/Dude.java
@@ -1,3 +1,10 @@
+import command.Command;
+import exception.DudeException;
+import parser.Parser;
+import storage.Storage;
+import storage.TaskList;
+import ui.Ui;
+
 import java.io.IOException;
 
 /**
