+++ src/main/java/command/CommandType.java
@@ -1,3 +1,5 @@
+package command;
+
 public enum CommandType {
     BYE,
     LIST,
