+++ src/main/java/exception/DudeException.java
@@ -1,9 +1,11 @@
+package exception;
+
 /**
  * Custom exception class for Dude-related exceptions.
  */
 public class DudeException extends Exception {
     /**
-     * The constructor creates a new DudeException object.
+     * The constructor creates a new exception.DudeException object.
      * 
      * @param message The error message to be displayed.
      */
